Upload of all work done during HTV 8
This is all work done during Hack the Valley 8 (during the fall of 2023) where I collaborated with an amazing group of people consisting of Min, Melanie and Rahma(and obviously Omer). We worked on my passion project and made it a reality. It was super fun working along y'all and I hope for the very best to you all :)